requirements.txt
Django==4.2.3
psycopg2
gunicorn
PyJwt
pytest-django
redis
fakeredis
djangorestframework
celery
watchfiles
django-cursor-pagination